fix: 폴더 이름 game-builder로 변경
