+++ src/game-builder/choice/applications/controllers/dto/delete-choice.dto.ts
