+++ src/game-builder/choice/domain/entities/page-choice.entity.ts
