+++ src/game-builder/game/application/controllers/dto/update-game.dto.ts
