+++ src/game-builder/game/domain/repositories/game.stub.repository.ts
