+++ src/game-builder/page/application/controllers/dto/check-spelling-by-external-service.dto.ts
