+++ src/game-builder/page/application/services/page.service.interface.ts
