+++ src/game-builder/page/application/services/page.stub.service.ts
