+++ src/game-builder/page/domain/entities/create-page.entity.ts
@@ -1,0 +1,14 @@
+export class CreatePageDomainEntity {
+  public createdAt: Date;
+  public updatedAt: Date;
+
+  constructor(
+    public content: string,
+    public abridgement: string,
+    public gameId: number,
+    public isEnding: boolean,
+  ) {
+    this.createdAt = new Date();
+    this.updatedAt = new Date();
+  }
+}
