+++ src/game-builder/page/domain/entities/page.entity.ts
