+++ src/game-builder/page/domain/entities/page.stub.entity.ts
