--- src/novel-game/choice/domain/entities/choice.entity.ts
@@ -1,12 +1,0 @@
-export class ChoiceDomainEntity {
-  constructor(
-    public id: number | null,
-    public title: string,
-    public description: string,
-    public parentPageId: number,
-    public childPageId: number | undefined,
-    public order: number,
-    public createdAt: Date,
-    public updatedAt: Date,
-  ) {}
-}
