--- src/novel-game/page/domain/repositories/page.repository.interface.ts
@@ -1,13 +1,0 @@
-import { Prisma } from '@prisma/client';
-import { PageDomainEntity } from '../entities/page.entity';
-
-export interface IPageRepository {
-  getOneById(
-    id: number,
-    transaction?: Prisma.TransactionClient,
-  ): Promise<PageDomainEntity | null>;
-  create(
-    page: PageDomainEntity,
-    transaction: Prisma.TransactionClient | undefined,
-  ): Promise<PageDomainEntity>;
-}
